requirements.txt
Django==3.2.2
djangorestframework==3.12.4
django-filter==2.4.0
psycopg2-binary>=2.8
ipython==8.8.0
python-decouple==3.6
django-cors-headers==3.10.0
Pillow==8.2.0
pytesseract==0.3.8
requests==2.25.1
opencv-python-headless==4.5.3.56
numpy==1.21.0